btrfs: add missing btrfs namespace prefix to tracepoints
A number of btrfs tracepoints are not properly namespaced:
  - __extent_writepage,
  - add_delayed_tree_ref,
  - run_delayed_tree_ref,
  - add_delayed_data_ref,
  - run_delayed_data_ref,
  - add_delayed_ref_head,
  - run_delayed_ref_head,
  - find_free_extent,
  - alloc_extent_state,
  - free_extent_state,
  - qgroup_num_dirty_extents,
  - qgroup_update_counters,
  - qgroup_update_reserve,
  - qgroup_meta_reserve,
  - qgroup_meta_convert,
  - qgroup_meta_free_all_pertrans.

This patch namespaces those tracepoints with the "btrfs" prefix to
follow the guidelines expressed in tracepoint.h.

This patch changes the names exposed to perf and ftrace which are
reflected in their produced traces and tracefs. As far as I know this
doesn't break any tool, but if anyone is aware of a tool relying on
those specific tracepoint names, please let us know so we can plan
a smooth transition.

Link: https://lore.kernel.org/lkml/1413560709-18704-1-git-send-email-jeremie.galarneau@efficios.com/
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Chris Mason <clm@fb.com>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: linux-btrfs@vger.kernel.org
Cc: Josef Bacik <josef@toxicpanda.com>
Cc: David Sterba <dsterba@suse.com>